changing from beta to prod ready
